Move Add Word Test to About page
